Moved cdogs delegate and models to common
