+++ Apps/Common/src/Delegates/ICDogsDelegate.cs
@@ -13,31 +13,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.WebClient.Models
+namespace HealthGateway.Common.Delegates
 {
-    using System.Text.Json.Serialization;
+    using System.Threading.Tasks;
+    using HealthGateway.Common.Models;
+    using HealthGateway.Common.Models.CDogs;
 
     /// <summary>
-    /// Model that defines the cdogs template.
+    /// The CDogs report service.
     /// </summary>
-    public class CDogsTemplateModel
+    public interface ICDogsDelegate
     {
         /// <summary>
-        /// Gets or sets the template content.
+        /// Generates a report based on the request model provided.
         /// </summary>
-        [JsonPropertyName("content")]
-        public string Content { get; set; } = string.Empty;
-
-        /// <summary>
-        /// Gets or sets the content encoding type.
-        /// </summary>
-        [JsonPropertyName("encodingType")]
-        public string EncodingType { get; set; } = "base64";
-
-        /// <summary>
-        /// Gets or sets the file type.
-        /// </summary>
-        [JsonPropertyName("fileType")]
-        public string FileType { get; set; } = "docx";
+        /// <param name="request">The cdogs request model.</param>
+        /// <returns>The report data.</returns>
+        Task<RequestResult<ReportModel>> GenerateReportAsync(CDogsRequestModel request);
     }
 }
