+++ Apps/Common/src/Models/CDogs/CDogsOptionsModel.cs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.WebClient.Models
+namespace HealthGateway.Common.Models.CDogs
 {
     using System.Text.Json.Serialization;
