+++ Apps/WebClient/src/Server/Services/IReportService.cs
@@ -16,6 +16,7 @@
 namespace HealthGateway.WebClient.Services
 {
     using HealthGateway.Common.Models;
+    using HealthGateway.Common.Models.CDogs;
     using HealthGateway.WebClient.Models;
 
     /// <summary>
