+++ Apps/WebClient/src/WebClient.xml
@@ -513,34 +513,6 @@
             <returns>An empty response.</returns>
             <response code="204">Webhook request received.</response>
         </member>
-        <member name="T:HealthGateway.WebClient.Delegates.CDogsDelegate">
-            <summary>
-            The CDogs report generator delegate.
-            </summary>
-        </member>
-        <member name="M:HealthGateway.WebClient.Delegates.CDogsDelegate.#ctor(Microsoft.Extensions.Logging.ILogger{HealthGateway.WebClient.Delegates.CDogsDelegate},HealthGateway.Common.Services.IHttpClientService,Microsoft.Extensions.Configuration.IConfiguration)">
-            <summary>
-            Initializes a new instance of the <see cref="T:HealthGateway.WebClient.Delegates.CDogsDelegate"/> class.
-            </summary>
-            <param name="logger">Injected Logger Provider.</param>
-            <param name="httpClientService">The injected http client service.</param>
-            <param name="configuration">The injected configuration provider.</param>
-        </member>
-        <member name="M:HealthGateway.WebClient.Delegates.CDogsDelegate.GenerateReportAsync(HealthGateway.WebClient.Models.CDogsRequestModel)">
-            <inheritdoc/>
-        </member>
-        <member name="T:HealthGateway.WebClient.Delegates.ICDogsDelegate">
-            <summary>
-            The CDogs report service.
-            </summary>
-        </member>
-        <member name="M:HealthGateway.WebClient.Delegates.ICDogsDelegate.GenerateReportAsync(HealthGateway.WebClient.Models.CDogsRequestModel)">
-            <summary>
-            Generates a report based on the request model provided.
-            </summary>
-            <param name="request">The cdogs request model.</param>
-            <returns>The report data.</returns>
-        </member>
         <member name="T:HealthGateway.WebClient.Listeners.BannerListener">
             <summary>
             Implements the abstract DB Listener and listens on the BannerChange channel.
@@ -714,98 +686,6 @@
         <member name="P:HealthGateway.WebClient.Models.BannerChangeEvent.Data">
             <summary>
             Gets or sets the Communication object.
-            </summary>
-        </member>
-        <member name="T:HealthGateway.WebClient.Models.CDogsConfig">
-            <summary>
-            The configuration for the ODR portion of the ResetMedStatement delegate.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsConfig.ServiceName">
-            <summary>
-            Gets or sets the OpenShift service name.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsConfig.ServiceHostSuffix">
-            <summary>
-            Gets or sets the host suffix used to lookup the service host.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsConfig.ServicePortSuffix">
-            <summary>
-            Gets or sets the port suffix used to lookup the service port.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsConfig.BaseEndpoint">
-            <summary>
-            Gets or sets the base Url used to connect to the ODR Proxy.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsConfig.DynamicServiceLookup">
-            <summary>
-            Gets or sets a value indicating whether dynamic service lookup should occur.
-            if enabled, the Url will be created using environment variables.
-            If not enabled, the configured Url will be used.
-            </summary>
-        </member>
-        <member name="T:HealthGateway.WebClient.Models.CDogsOptionsModel">
-            <summary>
-            Model that defines the cdogs options.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsOptionsModel.ConvertTo">
-            <summary>
-            Gets or sets the file type to convert to.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsOptionsModel.Overwrite">
-            <summary>
-            Gets or sets a value indicating whether to overwrite the file type.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsOptionsModel.ReportName">
-            <summary>
-            Gets or sets the report name.
-            </summary>
-        </member>
-        <member name="T:HealthGateway.WebClient.Models.CDogsRequestModel">
-            <summary>
-            Object that defines the cdogs request for creating a report.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsRequestModel.Data">
-            <summary>
-            Gets or sets the Json data.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsRequestModel.Options">
-            <summary>
-            Gets or sets the options.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsRequestModel.Template">
-            <summary>
-            Gets or sets the template.
-            </summary>
-        </member>
-        <member name="T:HealthGateway.WebClient.Models.CDogsTemplateModel">
-            <summary>
-            Model that defines the cdogs template.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsTemplateModel.Content">
-            <summary>
-            Gets or sets the template content.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsTemplateModel.EncodingType">
-            <summary>
-            Gets or sets the content encoding type.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.CDogsTemplateModel.FileType">
-            <summary>
-            Gets or sets the file type.
             </summary>
         </member>
         <member name="T:HealthGateway.WebClient.Models.CreateNoteRequest">
@@ -1068,21 +948,6 @@
         <member name="P:HealthGateway.WebClient.Models.OpenIdConnectConfiguration.Callbacks">
             <summary>
             Gets or sets the Callback URIs.
-            </summary>
-        </member>
-        <member name="T:HealthGateway.WebClient.Models.ReportModel">
-            <summary>
-            Object that defines a report.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.ReportModel.FileName">
-            <summary>
-            Gets or sets the report type.
-            </summary>
-        </member>
-        <member name="P:HealthGateway.WebClient.Models.ReportModel.Data">
-            <summary>
-            Gets or sets the report data.
             </summary>
         </member>
         <member name="T:HealthGateway.WebClient.Models.ReportFormatType">
@@ -2239,7 +2104,7 @@
         <member name="T:HealthGateway.WebClient.Services.ReportService">
             <inheritdoc />
         </member>
-        <member name="M:HealthGateway.WebClient.Services.ReportService.#ctor(Microsoft.Extensions.Logging.ILogger{HealthGateway.WebClient.Services.ReportService},HealthGateway.WebClient.Delegates.ICDogsDelegate)">
+        <member name="M:HealthGateway.WebClient.Services.ReportService.#ctor(Microsoft.Extensions.Logging.ILogger{HealthGateway.WebClient.Services.ReportService},HealthGateway.Common.Delegates.ICDogsDelegate)">
             <summary>
             Initializes a new instance of the <see cref="T:HealthGateway.WebClient.Services.ReportService"/> class.
             </summary>
